django-whatsapp/requirements.txt
aioredis==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==3.0.3
channels-redis==3.2.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==3.0.2
distro==1.9.0
Django==3.2
django-crispy-forms==1.11.2
djangorestframework==3.12.4
git-filter-repo==2.45.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.5.0
msgpack==1.1.0
openai==1.46.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyOpenSSL==24.2.1
pytz==2024.2
service-identity==24.1.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
zope.interface==7.0.3
